pyproject.toml
[tool.poetry]
name = "kakashi"
version = "0.1.0"
description = "An Open source general purpose discord Bot written in hikari library using Python"
authors = ["sarthak-py <shiva02939@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
lightbulb-ext-filament = "^0.1.2"
aiosqlite = "^0.17.0"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
